change os memory provider numa config

+++ test/memspaces/memspace_helpers.hpp
@@ -36,7 +36,7 @@ struct numaNodesTest : ::umf_test::test {
         }
     }
 
-    std::vector<size_t> nodeIds;
+    std::vector<unsigned> nodeIds;
     unsigned long maxNodeId = 0;
 };
